refactor: split decisionFormatter tests into focused modules (#818)
- Split 493-line test into 3 focused test files
- Leverage existing test helpers for data factories
- formatting.test.ts (127 lines): Core formatting logic
- edgeCases.test.ts (93 lines): Error handling & edge cases
- adaptiveLevels.test.ts (100 lines): Adaptive formatting tests
- testHelpers.ts (143 lines): Reusable test data factories

Updated strategy doc with progress:
- 3 files refactored this session
- 604 lines eliminated through helpers
- ~37 files remaining over 300 lines

Related to #818
